scripts/requirements.txt
firebase-admin==6.5.0
google-analytics-data==0.18.0
